Based on DesigniteJava’s rebellious Hierarchy smell, the method removeSI from InMemo_SN_TX_SemanticTag has been pushed down to InMemo_SN_TX_PeerSemanticTag, InMemoTimeSemanticTag, InMemoSpatialSemanticTag
